requirements_all.txt
absl-py==0.4.0
asn1crypto==0.24.0
astor==0.7.1
backcall==0.1.0
bleach==2.1.4
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
cryptography==2.3.1
cycler==0.10.0
Cython==0.28.5
dask==0.18.2
data==0.4
decorator==4.3.0
docopt==0.6.2
easydict==1.8
entrypoints==0.2.3
filetype==1.0.1
Flask==1.0.2
Flask-Login==0.4.1
Flask-WTF==0.14.2
funcsigs==1.0.2
gast==0.2.0
gevent==1.2.2
graphviz==0.8.4
greenlet==0.4.14
grpcio==1.14.1
h5py==2.8.0
html5lib==1.0.1
idna==2.6
imgaug==0.2.6
IPy==0.83
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
lightgbm==2.2.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mecab-python3==0.996.1
metrics==0.3.3
mistune==0.8.3
mmdnn==0.2.3
mudicom==0.1.2
mxnet-cu90==1.2.1.post1
nbconvert==5.3.1
nbformat==4.4.0
netaddr==0.7.19
networkx==2.1
notebook==5.6.0
numpy==1.14.5
opencv-python==3.4.2.17
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pathlib2==2.3.2
pathspec==0.5.5
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pipreqs==0.4.9
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.7
ptyprocess==0.6.0
pycparser==2.18
pydicom==1.1.0
Pygments==2.2.0
PyMySQL==0.9.2
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
rarfile==3.0
redis==2.10.6
requests==2.18.4
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
SimpleITK==1.1.0
six==1.11.0
SQLAlchemy==1.2.11
tensorboard==1.12.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tornado==5.1
tqdm==4.25.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.0
WTForms==2.2.1
xgboost==0.80
xlrd==1.1.0
xlwt==1.3.0
yarg==0.1.9
